Add blkx discard
This program helps to issue discard requests to ranges that are
not being written to by the regular blkx program. It's
blkx (fsx) reduced to discard request only.

The range of discard has lower bound of 20G, whereas writes has
upper bound of 10G, leaving a safe area of 10G between them,
yet the programs are full of asserts to make sure we're not
crossing the reserved range.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>